+++ src/onegov.notice/onegov/notice/upgrade.py
@@ -1,40 +1,19 @@
 from onegov.core.upgrade import upgrade_task
-from sqlalchemy import Enum
+from sqlalchemy import Column
+from sqlalchemy import Text
+from onegov.notice.models import OfficialNotice
 
 
-@upgrade_task('Adds accepted state')
-def add_accepted_state(context):
-    # Create a temporary definition and use it
-    tmp_type = Enum(
-        'drafted', 'submitted', 'published', 'rejected', 'accepted',
-        name='_official_notice_state'
-    )
-    tmp_type.create(context.operations.get_bind(), checkfirst=False)
-    context.operations.execute(
-        'ALTER TABLE official_notices ALTER COLUMN state '
-        'TYPE _official_notice_state '
-        'USING state::text::_official_notice_state'
-    )
+@upgrade_task('Adds a text column to the official notices')
+def add_text_column(context):
+    if not context.has_column('official_notices', 'text'):
+        # Create the new row
+        context.operations.add_column(
+            'official_notices',
+            Column('text', Text, nullable=True)
+        )
 
-    # Drop the old definition
-    old_type = Enum(
-        'drafted', 'submitted', 'published', 'rejected',
-        name='official_notice_state'
-    )
-    old_type.drop(context.operations.get_bind(), checkfirst=False)
-
-    # Create the new definition and use it
-    new_type = Enum(
-        'drafted', 'submitted', 'published', 'rejected', 'accepted',
-        name='official_notice_state'
-    )
-    new_type.create(context.operations.get_bind(), checkfirst=False)
-
-    context.operations.execute(
-        'ALTER TABLE official_notices ALTER COLUMN state '
-        'TYPE official_notice_state '
-        'USING state::text::official_notice_state'
-    )
-
-    # Drop the temporary defintion
-    tmp_type.drop(context.operations.get_bind(), checkfirst=False)
+        # Copy the data
+        for notice in context.session.query(OfficialNotice):
+            notice.text = notice.content.get('text', '')
+            notice.content = {}