+++ src/onegov.notice/HISTORY.rst
@@ -1,7 +1,7 @@
 Changelog
 ---------
 
-- Adds an 'accepted' state.
+- Allows filter notices by a search term.
   [msom]
 
 0.0.1 (2017-07-14)